+++ jsx/interfaces/FileInfoResponse.ts
@@ -1,0 +1,5 @@
+import {FileInfo} from '@uploadcare/react-widget';
+
+export default interface FileInfoResponse extends FileInfo {
+  attach_id: number;
+}