+++ jsx/interfaces/WpMediaModel.ts
@@ -1,0 +1,8 @@
+import WpMedia from './WpMedia';
+
+export default interface WpMediaModel {
+    cid: string;
+    attributes: WpMedia;
+    _listenId: string;
+    _pending: boolean;
+}