+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based now on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.0.3] — 2021-02-16
+
+* Fixed upload button position on "Add media" page.
+* Refactored code for better performance and control.
+* Now you can use Uploadcare Image Editor to modify images from Media Library.
+
 ## [3.0.2] — 2021-01-13
 
 * Plugin as class.